pyproject.toml
[build-system]
requires = ["setuptools", "setuptools-scm"]
build-backend = "setuptools.build_meta"

[project]
name = "makani"
authors = [
    { name="Thorsten Kurth", email = "tkurth@nvidia.com" },
    { name="Boris Bonev", email = "bbonev@nvidia.com" },
    { name="Jean Kossaifi", email = "jkossaifi@nvidia.com" },
    { name="Animashree Anandkumar" },
    { name="Kamyar Azizzadenesheli" },
    { name="Noah Brenowitz" },
    { name="Ashesh Chattopadhyay" },
    { name="Yair Cohen" },
    { name="David Hall" },
    { name="Peter Harrington" },
    { name="Pedram Hassanzadeh" },
    { name="Christian Hundt" },
    { name="Karthik Kashinath" },
    { name="Zongyi Li" },
    { name="Morteza Mardani" },
    { name="Jaideep Pathak" },
    { name="Mike Pritchard" },
    { name="David Pruitt" },
    { name="Sanjeev Raja" },
    { name="Shashank Subramanian" },
]

maintainers = [
    { name="Thorsten Kurth", email = "tkurth@nvidia.com" },
    { name="Boris Bonev", email = "bbonev@nvidia.com" },
]

description = "Massively parallel training of machine-learning based weather and climate models"
readme = "README.md"
requires-python = ">=3.10"

classifiers = [
    "Development Status :: Alpha",
    "Programming Language :: Python :: 3.10",
    "License :: OSI Approved :: Apache Software License",
    "Operating System :: OS Independent",
]

dynamic = ["version"]

dependencies = [
    "torch>=2.0.0",
    "numpy>=1.22.4,<1.25",
    "nvidia_dali_cuda110>=1.16.0",
    "nvidia-modulus>=0.5.0a0"
]

[tool.setuptools.dynamic]
version = {attr = "makani.__version__"}

[tool.setuptools.packages.find]
    include = ["makani*"]

[project.optional-dependencies]
dev = [
    "pytest>=6.0.0",
    "black>=22.10.0",
    "coverage>=6.5.0",
]

all = [
    "wandb>=0.13.7",
    "imageio>=2.28.1",
    "moviepy>=1.0.3",
    "tqdm>=4.60.0",
]

[tool.pytest.ini_options]
minversion = "6.0"
addopts = ""
testpaths = [
    "tests",
    "makani",
]

[tool.black]
line-length = 120
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''